requirements_test_all.txt
# Home Assistant test
# linters such as flake8 and pylint should be pinned, as new releases
# make new things fail. Manually update these pins when pulling in a
# new version
flake8==3.3
pylint==1.6.5
mypy==0.550
pydocstyle==1.1.1
coveralls>=1.1
pytest>=2.9.2
pytest-aiohttp>=0.1.3
pytest-cov>=2.3.1
pytest-timeout>=1.2.0
pytest-catchlog>=1.2.2
pytest-sugar>=0.7.1
requests_mock>=1.0
mock-open>=1.3.1
flake8-docstrings==1.0.2
asynctest>=0.8.0
freezegun>=0.3.8


# homeassistant.components.notify.html5
PyJWT==1.5.3

# homeassistant.components.media_player.sonos
SoCo==0.13

# homeassistant.components.device_tracker.automatic
aioautomatic==0.6.4

# homeassistant.components.emulated_hue
# homeassistant.components.http
aiohttp_cors==0.5.3

# homeassistant.components.notify.apns
apns2==0.3.0

# homeassistant.components.calendar.caldav
caldav==0.5.0

# homeassistant.components.sensor.coinmarketcap
coinmarketcap==4.1.1

# homeassistant.components.device_tracker.upc_connect
defusedxml==0.5.0

# homeassistant.components.sensor.dsmr
dsmr_parser==0.11

# homeassistant.components.sensor.season
ephem==3.7.6.0

# homeassistant.components.climate.honeywell
evohomeclient==0.2.5

# homeassistant.components.feedreader
# homeassistant.components.sensor.geo_rss_events
feedparser==5.2.1

# homeassistant.components.conversation
fuzzywuzzy==0.15.1

# homeassistant.components.tts.google
gTTS-token==1.1.1

# homeassistant.components.ffmpeg
ha-ffmpeg==1.9

# homeassistant.components.sensor.geo_rss_events
haversine==0.4.5

# homeassistant.components.mqtt.server
hbmqtt==0.9.1

# homeassistant.components.binary_sensor.workday
holidays==0.8.1

# homeassistant.components.frontend
home-assistant-frontend==20171216.0

# homeassistant.components.influxdb
# homeassistant.components.sensor.influxdb
influxdb==4.1.1

# homeassistant.components.dyson
libpurecoollink==0.4.2

# homeassistant.components.media_player.soundtouch
libsoundtouch==0.7.2

# homeassistant.components.sensor.mfi
# homeassistant.components.switch.mfi
mficlient==0.3.0

# homeassistant.components.binary_sensor.trend
# homeassistant.components.image_processing.opencv
numpy==1.13.3

# homeassistant.components.mqtt
# homeassistant.components.shiftr
paho-mqtt==1.3.1

# homeassistant.components.device_tracker.aruba
# homeassistant.components.device_tracker.asuswrt
# homeassistant.components.device_tracker.cisco_ios
# homeassistant.components.device_tracker.unifi_direct
# homeassistant.components.media_player.pandora
pexpect==4.0.1

# homeassistant.components.pilight
pilight==0.1.1

# homeassistant.components.sensor.mhz19
# homeassistant.components.sensor.serial_pm
pmsensor==0.4

# homeassistant.components.prometheus
prometheus_client==0.0.21

# homeassistant.components.canary
py-canary==0.2.3

# homeassistant.components.zwave
pydispatcher==2.0.5

# homeassistant.components.litejet
pylitejet==0.1

# homeassistant.components.alarm_control_panel.nx584
# homeassistant.components.binary_sensor.nx584
pynx584==0.4

# homeassistant.components.sensor.darksky
python-forecastio==1.3.5

# homeassistant.components.sensor.whois
pythonwhois==2.4.3

# homeassistant.components.device_tracker.unifi
pyunifi==2.13

# homeassistant.components.notify.html5
pywebpush==1.3.0

# homeassistant.components.python_script
restrictedpython==4.0b2

# homeassistant.components.rflink
rflink==0.0.34

# homeassistant.components.ring
ring_doorbell==0.1.9

# homeassistant.components.media_player.yamaha
rxv==0.5.1

# homeassistant.components.sleepiq
sleepyq==0.6

# homeassistant.components.climate.honeywell
somecomfort==0.5.0

# homeassistant.components.recorder
# homeassistant.scripts.db_migrator
sqlalchemy==1.1.15

# homeassistant.components.statsd
statsd==3.2.1

# homeassistant.components.camera.uvc
uvcclient==0.10.1

# homeassistant.components.vultr
vultr==0.1.2

# homeassistant.components.wake_on_lan
# homeassistant.components.media_player.panasonic_viera
# homeassistant.components.media_player.samsungtv
# homeassistant.components.media_player.webostv
# homeassistant.components.switch.wake_on_lan
wakeonlan==0.2.2

# homeassistant.components.cloud
warrant==0.6.1

# homeassistant.components.sensor.yahoo_finance
yahoo-finance==1.4.0
